Distance calculation added
- replaced fake repositories to the real ones
- updated tests
- added docker-compose and Makefile to run infrastruction

+++ users/pkg/location/location.go
@@ -2,6 +2,6 @@ package location
 
 type Location struct {
 	Username  string  `json:"username"`
-	Longitude float32 `json:"longitude"`
-	Latitude  float32 `json:"latitude"`
+	Longitude float64 `json:"longitude"`
+	Latitude  float64 `json:"latitude"`
 }
